Feature: Implement bucket system for active cover expiry

+++ contracts/mocks/MCR/MCRMockCover.sol
@@ -6,13 +6,13 @@ import "../../interfaces/ICover.sol";
 
 contract MCRMockCover {
 
-  mapping(uint24 => uint) public totalActiveCoverInAsset;
+  mapping(uint => uint) public totalActiveCoverInAsset;
 
-  function getTotalActiveCoverInAsset(uint24 coverAsset) external view returns (uint) {
+  function getTotalActiveCoverInAsset(uint coverAsset) external view returns (uint) {
     return totalActiveCoverInAsset[coverAsset];
   }
 
-  function setTotalActiveCoverInAsset(uint24 asset, uint amount) public {
+  function setTotalActiveCoverInAsset(uint asset, uint amount) public {
     totalActiveCoverInAsset[asset] = amount;
   }
 }
